Inserção do Questionario inicial, correção da responsividade dos formularios
