--- src/components/container/questions/index.tsx
@@ -1,8 +1,0 @@
-
-export default function Index(){
-  return(
-    <>
-    <iframe src="https://docs.google.com/forms/d/e/1FAIpQLSefdMYNtKCdNfa6GDvzOUl9IOpRe682asL_3dDVwaRBJ0z5pA/viewform?embedded=true" width="100%" height="500px" >Carregando…</iframe>
-      </>
-  )
-}
