+++ src/core/enums.ts
@@ -6,6 +6,7 @@ export enum routerEnum {
 	HOME = '/home',
 	FORM = '/form',
 	FORM_ANSWER = '/form-answer',
+	QUESTION ='/question',
 }
 
 export enum localStorageKeyEnum {
