+++ src/pages/form/style.ts
@@ -2,15 +2,15 @@ import { CSSProperties } from 'react';
 import { theme } from 'src/core/theme';
 
 export const mainContainerStyle: CSSProperties = {
-	display: 'grid',
-	gridTemplateColumns: '20% 20% 20% 20% 20%',
-	gridTemplateRows: 'auto',
+	display: 'flex',
+	flexWrap: 'wrap'
 };
 
 export const formButtonStyle: CSSProperties = {
 	...theme.button,
 	backgroundColor: theme.primaryColor,
 	color: theme.white,
+	width: '200px',
 	margin: '1rem',
   height: 'auto',
 	borderWidth: '0.5rem',
